Create & Display Product
Creating an Ad in `Sell` page and displaying it in the `profile`. Also in `Register.js` and `Login.js`, the `navigate` function has an argument `{replace : true}`. This means that when the user navigates to the new page, there will be no record of the prev page in the history stack. Therefore, if they click the back button, they won't go back to the `login` page.